requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1
Pillow==10.1.0
python-multipart==0.0.6 